+++ lua/plugins/null-ls.lua
@@ -17,13 +17,19 @@ return {
         debug = false,
         sources = {
           formatting.prettier.with({
+            -- milliseconds
+            timeout = 5000,
             extra_args = { "--single-quote", "false" },
           }),
-          formatting.stylua.with({ extra_args = { "--indent-type", "Spaces", "--indent-width", "2" } }),
+          formatting.stylua.with({
+            timeout = 5000,
+            extra_args = { "--indent-type", "Spaces", "--indent-width", "2" },
+          }),
           formatting.terraform_fmt,
           formatting.goimports,
           formatting.gofumpt,
           formatting.latexindent.with({
+            timeout = 5000,
             extra_args = { "-g", "/dev/null" }, -- https://github.com/cmhughes/latexindent.pl/releases/tag/V3.9.3
           }),
           actions.shellcheck,
@@ -32,10 +38,19 @@ return {
           diagnostics.ruff,
           formatting.shellharden,
           formatting.google_java_format,
-          formatting.black.with({ extra_args = { "--fast" } }),
-          formatting.sql_formatter.with({ extra_args = { "--config" } }),
+          formatting.black.with({
+            timeout = 5000,
+            extra_args = { "--fast" },
+          }),
+          formatting.sql_formatter.with({
+            timeout = 5000,
+            extra_args = { "--config" },
+          }),
           formatting.markdownlint,
-          formatting.beautysh.with({ extra_args = { "--indent-size", "2" } }),
+          formatting.beautysh.with({
+            timeout = 5000,
+            extra_args = { "--indent-size", "2" },
+          }),
           diagnostics.zsh.with({
             filetypes = { "zsh" },
           }),