+++ scripts/install_neovim.sh
@@ -12,19 +12,19 @@ DOC_HOMEBREW="https://docs.brew.sh"
 BREW_EXE="brew"
 export PATH=${HOME}/.local/bin:${PATH}
 
-abort () {
+abort() {
   printf "\nERROR: %s\n" "$@" >&2
   exit 1
 }
 
-log () {
-  [ "${quiet}" ] || {
+log() {
+  [ "$quiet" ] || {
     printf "\n\t%s" "$@"
   }
 }
 
-check_prerequisites () {
-  if [ -z "${BASH_VERSION:-}" ]; then
+check_prerequisites() {
+  if [ "${BASH_VERSION:-}" = "" ]; then
     abort "Bash is required to interpret this script."
   fi
 
@@ -38,49 +38,44 @@ check_prerequisites () {
   fi
 }
 
-install_homebrew () {
+install_homebrew() {
   if ! command -v brew >/dev/null 2>&1; then
-    [ "${debug}" ] && START_SECONDS=$(date +%s)
+    [ "$debug" ] && START_SECONDS=$(date +%s)
     log "Installing Homebrew ..."
     BREW_URL="https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh"
     have_curl=$(type -p curl)
-    [ "${have_curl}" ] || abort "The curl command could not be located."
-    curl -fsSL "${BREW_URL}" > /tmp/brew-$$.sh
+    [ "$have_curl" ] || abort "The curl command could not be located."
+    curl -fsSL "$BREW_URL" >/tmp/brew-$$.sh
     [ $? -eq 0 ] || {
       rm -f /tmp/brew-$$.sh
-      curl -kfsSL "${BREW_URL}" > /tmp/brew-$$.sh
+      curl -kfsSL "$BREW_URL" >/tmp/brew-$$.sh
     }
     [ -f /tmp/brew-$$.sh ] || abort "Brew install script download failed"
     chmod 755 /tmp/brew-$$.sh
-    NONINTERACTIVE=1 /bin/bash -c "/tmp/brew-$$.sh" > /dev/null 2>&1
+    NONINTERACTIVE=1 /bin/bash -c "/tmp/brew-$$.sh" >/dev/null 2>&1
     rm -f /tmp/brew-$$.sh
     export HOMEBREW_NO_INSTALL_CLEANUP=1
     export HOMEBREW_NO_ENV_HINTS=1
     export HOMEBREW_NO_AUTO_UPDATE=1
-    [ "${quiet}" ] || printf " done"
-    if [ -f ${HOME}/.profile ]
-    then
+    [ "$quiet" ] || printf " done"
+    if [ -f "$HOME"/.profile ]; then
       BASHINIT="${HOME}/.profile"
     else
-      if [ -f ${HOME}/.bashrc ]
-      then
+      if [ -f "$HOME"/.bashrc ]; then
         BASHINIT="${HOME}/.bashrc"
       else
         BASHINIT="${HOME}/.profile"
       fi
     fi
-    if [ -x /home/linuxbrew/.linuxbrew/bin/brew ]
-    then
+    if [ -x /home/linuxbrew/.linuxbrew/bin/brew ]; then
       HOMEBREW_HOME="/home/linuxbrew/.linuxbrew"
       BREW_EXE="${HOMEBREW_HOME}/bin/brew"
     else
-      if [ -x /usr/local/bin/brew ]
-      then
+      if [ -x /usr/local/bin/brew ]; then
         HOMEBREW_HOME="/usr/local"
         BREW_EXE="${HOMEBREW_HOME}/bin/brew"
       else
-        if [ -x /opt/homebrew/bin/brew ]
-        then
+        if [ -x /opt/homebrew/bin/brew ]; then
           HOMEBREW_HOME="/opt/homebrew"
           BREW_EXE="${HOMEBREW_HOME}/bin/brew"
         else
@@ -89,52 +84,50 @@ install_homebrew () {
       fi
     fi
 
-    if [ -f "${BASHINIT}" ]
-    then
-      grep "eval \"\$(${BREW_EXE} shellenv)\"" "${BASHINIT}" > /dev/null || {
-        echo 'if [ -x XXX ]; then' | sed -e "s%XXX%${BREW_EXE}%" >> "${BASHINIT}"
-        echo '  eval "$(XXX shellenv)"' | sed -e "s%XXX%${BREW_EXE}%" >> "${BASHINIT}"
-        echo 'fi' >> "${BASHINIT}"
+    if [ -f "$BASHINIT" ]; then
+      grep "eval \"\$(${BREW_EXE} shellenv)\"" "$BASHINIT" >/dev/null || {
+        echo 'if [ -x XXX ]; then' | sed -e "s%XXX%${BREW_EXE}%" >>"$BASHINIT"
+        echo '  eval "$(XXX shellenv)"' | sed -e "s%XXX%${BREW_EXE}%" >>"$BASHINIT"
+        echo 'fi' >>"$BASHINIT"
       }
-      grep "eval \"\$(zoxide init" "${BASHINIT}" > /dev/null || {
-        echo 'if command -v zoxide > /dev/null; then' >> "${BASHINIT}"
-        echo '  eval "$(zoxide init bash)"' >> "${BASHINIT}"
-        echo 'fi' >> "${BASHINIT}"
+      grep "eval \"\$(zoxide init" "$BASHINIT" >/dev/null || {
+        echo 'if command -v zoxide > /dev/null; then' >>"$BASHINIT"
+        echo '  eval "$(zoxide init bash)"' >>"$BASHINIT"
+        echo 'fi' >>"$BASHINIT"
       }
     else
-      echo 'if [ -x XXX ]; then' | sed -e "s%XXX%${BREW_EXE}%" > "${BASHINIT}"
-      echo '  eval "$(XXX shellenv)"' | sed -e "s%XXX%${BREW_EXE}%" >> "${BASHINIT}"
-      echo 'fi' >> "${BASHINIT}"
-      echo 'if command -v zoxide > /dev/null; then' >> "${BASHINIT}"
-      echo '  eval "$(zoxide init bash)"' >> "${BASHINIT}"
-      echo 'fi' >> "${BASHINIT}"
+      echo 'if [ -x XXX ]; then' | sed -e "s%XXX%${BREW_EXE}%" >"$BASHINIT"
+      echo '  eval "$(XXX shellenv)"' | sed -e "s%XXX%${BREW_EXE}%" >>"$BASHINIT"
+      echo 'fi' >>"$BASHINIT"
+      echo 'if command -v zoxide > /dev/null; then' >>"$BASHINIT"
+      echo '  eval "$(zoxide init bash)"' >>"$BASHINIT"
+      echo 'fi' >>"$BASHINIT"
     fi
     [ -f "${HOME}/.zshrc" ] && {
-      grep "eval \"\$(${BREW_EXE} shellenv)\"" "${HOME}/.zshrc" > /dev/null || {
-        echo 'if [ -x XXX ]; then' | sed -e "s%XXX%${BREW_EXE}%" >> "${HOME}/.zshrc"
-        echo '  eval "$(XXX shellenv)"' | sed -e "s%XXX%${BREW_EXE}%" >> "${HOME}/.zshrc"
-        echo 'fi' >> "${HOME}/.zshrc"
+      grep "eval \"\$(${BREW_EXE} shellenv)\"" "${HOME}/.zshrc" >/dev/null || {
+        echo 'if [ -x XXX ]; then' | sed -e "s%XXX%${BREW_EXE}%" >>"${HOME}/.zshrc"
+        echo '  eval "$(XXX shellenv)"' | sed -e "s%XXX%${BREW_EXE}%" >>"${HOME}/.zshrc"
+        echo 'fi' >>"${HOME}/.zshrc"
       }
-      grep "eval \"\$(zoxide init" "${HOME}/.zshrc" > /dev/null || {
-        echo 'if command -v zoxide > /dev/null; then' >> "${HOME}/.zshrc"
-        echo '  eval "$(zoxide init zsh)"' >> "${HOME}/.zshrc"
-        echo 'fi' >> "${HOME}/.zshrc"
+      grep "eval \"\$(zoxide init" "${HOME}/.zshrc" >/dev/null || {
+        echo 'if command -v zoxide > /dev/null; then' >>"${HOME}/.zshrc"
+        echo '  eval "$(zoxide init zsh)"' >>"${HOME}/.zshrc"
+        echo 'fi' >>"${HOME}/.zshrc"
       }
     }
-    eval "$(${BREW_EXE} shellenv)"
-    have_brew=`type -p brew`
-    [ "${have_brew}" ] && BREW_EXE="brew"
-    [ "${debug}" ] && {
+    eval "$("$BREW_EXE" shellenv)"
+    have_brew=$(type -p brew)
+    [ "$have_brew" ] && BREW_EXE="brew"
+    [ "$debug" ] && {
       FINISH_SECONDS=$(date +%s)
-      ELAPSECS=$(( FINISH_SECONDS - START_SECONDS ))
-      ELAPSED=`eval "echo $(date -ud "@$ELAPSECS" +'$((%s/3600/24)) days %H hr %M min %S sec')"`
+      ELAPSECS=$((FINISH_SECONDS - START_SECONDS))
+      ELAPSED=$(eval "echo $(date -ud "@$ELAPSECS" +'$((%s/3600/24)) days %H hr %M min %S sec')")
       printf "\nHomebrew install elapsed time = ${ELAPSED}\n"
     }
-    [ "${HOMEBREW_HOME}" ] || {
+    [ "$HOMEBREW_HOME" ] || {
       brewpath=$(command -v brew)
-      if [ $? -eq 0 ]
-      then
-        HOMEBREW_HOME=`dirname ${brewpath} | sed -e "s%/bin$%%"`
+      if [ $? -eq 0 ]; then
+        HOMEBREW_HOME=$(dirname "$brewpath" | sed -e "s%/bin$%%")
       else
         HOMEBREW_HOME="Unknown"
       fi
@@ -145,72 +138,69 @@ install_homebrew () {
 }
 
 brew_install() {
-	brewpkg="$1"
-  if command -v ${brewpkg} >/dev/null 2>&1
-	then
+  brewpkg="$1"
+  if command -v "$brewpkg" >/dev/null 2>&1; then
     log "Using previously installed ${brewpkg} ..."
-	else
+  else
     log "Installing ${brewpkg} ..."
-    [ "${debug}" ] && START_SECONDS=$(date +%s)
-    ${BREW_EXE} install --quiet ${brewpkg} > /dev/null 2>&1
-    [ $? -eq 0 ] || ${BREW_EXE} link --overwrite --quiet ${pkg} > /dev/null 2>&1
-    if [ "${debug}" ]
-    then
+    [ "$debug" ] && START_SECONDS=$(date +%s)
+    "$BREW_EXE" install --quiet "$brewpkg" >/dev/null 2>&1
+    [ $? -eq 0 ] || "$BREW_EXE" link --overwrite --quiet "$pkg" >/dev/null 2>&1
+    if [ "$debug" ]; then
       FINISH_SECONDS=$(date +%s)
-      ELAPSECS=$(( FINISH_SECONDS - START_SECONDS ))
-      ELAPSED=`eval "echo $(date -ud "@$ELAPSECS" +'$((%s/3600/24)) days %H hr %M min %S sec')"`
+      ELAPSECS=$((FINISH_SECONDS - START_SECONDS))
+      ELAPSED=$(eval "echo $(date -ud "@$ELAPSECS" +'$((%s/3600/24)) days %H hr %M min %S sec')")
       printf " elapsed time = %s${ELAPSED}"
     fi
-	fi
-  [ "${quiet}" ] || printf " done"
+  fi
+  [ "$quiet" ] || printf " done"
 }
 
-install_neovim_dependencies () {
-  [ "${quiet}" ] || printf "\nInstalling dependencies"
+install_neovim_dependencies() {
+  [ "$quiet" ] || printf "\nInstalling dependencies"
   PKGS="git curl tar unzip lazygit fd fzf xclip zoxide"
-  for pkg in ${PKGS}
-  do
-    if command -v ${pkg} >/dev/null 2>&1
-	  then
+  for pkg in "$PKGS"; do
+    if command -v "$pkg" >/dev/null 2>&1; then
       log "Using previously installed ${pkg}"
-	  else
-		  brew_install "${pkg}"
+    else
+      brew_install "$pkg"
     fi
   done
-  if command -v rg >/dev/null 2>&1
-	then
+  if command -v rg >/dev/null 2>&1; then
     log "Using previously installed ripgrep"
-	else
-	  brew_install ripgrep
-	fi
-  [ "${quiet}" ] || printf "\n"
+  else
+    brew_install ripgrep
+  fi
+  if command -v ag >/dev/null 2>&1; then
+    log "Using previously installed the_silver_searcher"
+  else
+    brew_install the_silver_searcher
+  fi
+  [ "$quiet" ] || printf "\n"
 }
 
-install_neovim_head () {
-  ${BREW_EXE} link -q libuv > /dev/null 2>&1
+install_neovim_head() {
+  "$BREW_EXE" link -q libuv >/dev/null 2>&1
   log "Compiling and installing Neovim, please be patient ..."
-  if [ "${debug}" ]
-  then
+  if [ "$debug" ]; then
     START_SECONDS=$(date +%s)
-    ${BREW_EXE} install --HEAD neovim
+    "$BREW_EXE" install --HEAD neovim
   else
-    ${BREW_EXE} install -q --HEAD neovim > /dev/null 2>&1
+    "$BREW_EXE" install -q --HEAD neovim >/dev/null 2>&1
   fi
-  if [ "${debug}" ]
-  then
+  if [ "$debug" ]; then
     FINISH_SECONDS=$(date +%s)
-    ELAPSECS=$(( FINISH_SECONDS - START_SECONDS ))
-    ELAPSED=`eval "echo $(date -ud "@$ELAPSECS" +'$((%s/3600/24)) days %H hr %M min %S sec')"`
+    ELAPSECS=$((FINISH_SECONDS - START_SECONDS))
+    ELAPSED=$(eval "echo $(date -ud "@$ELAPSECS" +'$((%s/3600/24)) days %H hr %M min %S sec')")
     printf "\nInstall Neovim HEAD elapsed time = %s${ELAPSED}\n"
   fi
-  [ "${quiet}" ] || printf " done"
+  [ "$quiet" ] || printf " done"
 }
 
-check_python () {
+check_python() {
   brew_path=$(command -v brew)
-  brew_dir=$(dirname ${brew_path})
-  if [ -x ${brew_dir}/python3 ]
-  then
+  brew_dir=$(dirname "$brew_path")
+  if [ -x "$brew_dir"/python3 ]; then
     PYTHON="${brew_dir}/python3"
   else
     PYTHON=$(command -v python3)
@@ -218,17 +208,16 @@ check_python () {
 }
 
 # Brew doesn't create a python symlink so we do so here
-link_python () {
+link_python() {
   python3_path=$(command -v python3)
-  [ "${python3_path}" ] && {
-    python_dir=`dirname ${python3_path}`
-    if [ -w ${python_dir} ]
-    then
-      rm -f ${python_dir}/python
-      ln -s ${python_dir}/python3 ${python_dir}/python
+  [ "$python3_path" ] && {
+    python_dir=$(dirname "$python3_path")
+    if [ -w "$python_dir" ]; then
+      rm -f "$python_dir"/python
+      ln -s "$python_dir"/python3 "$python_dir"/python
     else
-      sudo rm -f ${python_dir}/python
-      sudo ln -s ${python_dir}/python3 ${python_dir}/python
+      sudo rm -f "$python_dir"/python
+      sudo ln -s "$python_dir"/python3 "$python_dir"/python
     fi
   }
 }
@@ -236,63 +225,62 @@ link_python () {
 # Language servers are mostly being handled by Mason
 # but some external utilities are required
 install_language_servers() {
-  [ "${quiet}" ] || printf "\nInstalling language servers and tools"
+  [ "$quiet" ] || printf "\nInstalling language servers and tools"
 
-	brew_install ccls
-  ${BREW_EXE} link --overwrite --quiet ccls > /dev/null 2>&1
+  brew_install ccls
+  "$BREW_EXE" link --overwrite --quiet ccls >/dev/null 2>&1
 
   for pkg in golangci-lint jdtls rust-analyzer taplo eslint terraform \
-             yarn julia composer php deno
-  do
-		brew_install "${pkg}"
+    yarn julia composer php deno; do
+    brew_install "$pkg"
   done
 
-  [ "${PYTHON}" ] && {
-    ${PYTHON} -m pip install python-lsp-server > /dev/null 2>&1
+  [ "$PYTHON" ] && {
+    "$PYTHON" -m pip install python-lsp-server >/dev/null 2>&1
   }
   if command -v go >/dev/null 2>&1; then
-    go install golang.org/x/tools/gopls@latest > /dev/null 2>&1
+    go install golang.org/x/tools/gopls@latest >/dev/null 2>&1
   fi
-  [ "${quiet}" ] || printf "\nDone"
+  [ "$quiet" ] || printf "\nDone"
 }
 
 install_tools() {
-  [ "${quiet}" ] || printf "\nInstalling Python dependencies"
+  [ "$quiet" ] || printf "\nInstalling Python dependencies"
   check_python
-  [ "${PYTHON}" ] || {
+  [ "$PYTHON" ] || {
     # Could not find Python, install with Homebrew
     log 'Installing Python with Homebrew ...'
-    ${BREW_EXE} install --quiet python > /dev/null 2>&1
-    [ $? -eq 0 ] || ${BREW_EXE} link --overwrite --quiet python > /dev/null 2>&1
+    "$BREW_EXE" install --quiet python >/dev/null 2>&1
+    [ $? -eq 0 ] || "$BREW_EXE" link --overwrite --quiet python >/dev/null 2>&1
     link_python
     check_python
-    [ "${quiet}" ] || printf " done"
+    [ "$quiet" ] || printf " done"
   }
-  [ "${PYTHON}" ] && {
+  [ "$PYTHON" ] && {
     log 'Upgrading pip, setuptools, wheel, doq, and pynvim ...'
-    ${PYTHON} -m pip install --upgrade pip > /dev/null 2>&1
-    ${PYTHON} -m pip install --upgrade setuptools > /dev/null 2>&1
-    ${PYTHON} -m pip install wheel > /dev/null 2>&1
-    ${PYTHON} -m pip install pynvim doq > /dev/null 2>&1
-    [ "${quiet}" ] || printf " done"
+    "$PYTHON" -m pip install --upgrade pip >/dev/null 2>&1
+    "$PYTHON" -m pip install --upgrade setuptools >/dev/null 2>&1
+    "$PYTHON" -m pip install wheel >/dev/null 2>&1
+    "$PYTHON" -m pip install pynvim doq >/dev/null 2>&1
+    [ "$quiet" ] || printf " done"
   }
-  [ "${quiet}" ] || printf "\nDone"
+  [ "$quiet" ] || printf "\nDone"
 
-  [ "${quiet}" ] || printf "\nInstalling npm, treesitter, and cargo dependencies"
+  [ "$quiet" ] || printf "\nInstalling npm, treesitter, and cargo dependencies"
   have_npm=$(type -p npm)
-  [ "${have_npm}" ] && {
+  [ "$have_npm" ] && {
     log "Installing Neovim npm package ..."
-    npm i -g neovim > /dev/null 2>&1
-    [ "${quiet}" ] || printf " done"
+    npm i -g neovim >/dev/null 2>&1
+    [ "$quiet" ] || printf " done"
 
     log "Installing the icon font for Visual Studio Code ..."
-    npm i -g @vscode/codicons > /dev/null 2>&1
-    [ "${quiet}" ] || printf " done"
+    npm i -g @vscode/codicons >/dev/null 2>&1
+    [ "$quiet" ] || printf " done"
   }
 
-	brew_install tree-sitter
+  brew_install tree-sitter
   if command -v tree-sitter >/dev/null 2>&1; then
-    tree-sitter init-config > /dev/null 2>&1
+    tree-sitter init-config >/dev/null 2>&1
   fi
 
   # if command -v cargo >/dev/null 2>&1; then
@@ -301,23 +289,22 @@ install_tools() {
 
   GHUC="https://raw.githubusercontent.com"
   JETB_URL="${GHUC}/JetBrains/JetBrainsMono/master/install_manual.sh"
-  [ "${quiet}" ] || printf "\n\tInstalling JetBrains Mono font ... "
-  curl -fsSL "${JETB_URL}" > /tmp/jetb-$$.sh
+  [ "$quiet" ] || printf "\n\tInstalling JetBrains Mono font ... "
+  curl -fsSL "$JETB_URL" >/tmp/jetb-$$.sh
   [ $? -eq 0 ] || {
     rm -f /tmp/jetb-$$.sh
-    curl -kfsSL "${JETB_URL}" > /tmp/jetb-$$.sh
+    curl -kfsSL "$JETB_URL" >/tmp/jetb-$$.sh
   }
   [ -f /tmp/jetb-$$.sh ] && {
     chmod 755 /tmp/jetb-$$.sh
-    /bin/bash -c "/tmp/jetb-$$.sh" > /dev/null 2>&1
+    /bin/bash -c "/tmp/jetb-$$.sh" >/dev/null 2>&1
     rm -f /tmp/jetb-$$.sh
   }
-  [ "${quiet}" ] || printf "done\nDone\n"
+  [ "$quiet" ] || printf "done\nDone\n"
 }
 
-main () {
-  if [ "${lang_tools}" ]
-  then
+main() {
+  if [ "$lang_tools" ]; then
     install_homebrew
     install_neovim_dependencies
     install_language_servers
@@ -326,9 +313,9 @@ main () {
     check_prerequisites
     if command -v nvim >/dev/null 2>&1; then
       nvim_version=$(nvim --version | head -1 | grep -o '[0-9]\.[0-9]')
-      if (( $(echo "$nvim_version < 0.9 " |bc -l) )); then
+      if (($(echo "$nvim_version < 0.9 " | bc -l))); then
         printf "\nCurrently installed Neovim is less than version 0.9"
-        [ "${nvim_head}" ] && {
+        [ "$nvim_head" ] && {
           install_homebrew
           install_neovim_dependencies
           printf "\nInstalling latest Neovim version with Homebrew"
@@ -339,8 +326,7 @@ main () {
       install_homebrew
       install_neovim_dependencies
       printf "\nNeovim not found, installing Neovim with Homebrew"
-      if [ "${nvim_head}" ]
-      then
+      if [ "$nvim_head" ]; then
         install_neovim_head
       else
         brew_install neovim
@@ -357,41 +343,39 @@ lang_tools=
 while getopts "dhlq" flag; do
   case $flag in
     d)
-        debug=1
-        ;;
+      debug=1
+      ;;
     h)
-        nvim_head=
-        ;;
+      nvim_head=
+      ;;
     l)
-        lang_tools=1
-        ;;
+      lang_tools=1
+      ;;
     q)
-        quiet=1
-        ;;
-    *)
-        ;;
+      quiet=1
+      ;;
+    *) ;;
   esac
 done
 
 currlimit=$(ulimit -n)
 hardlimit=$(ulimit -Hn)
-[ "${hardlimit}" == "unlimited" ] && hardlimit=9999
-if [ ${hardlimit} -gt 4096 ]
-then
+[ "$hardlimit" == "unlimited" ] && hardlimit=9999
+if [ "$hardlimit" -gt 4096 ]; then
   ulimit -n 4096
 else
-  ulimit -n ${hardlimit}
+  ulimit -n "$hardlimit"
 fi
 
-[ "${debug}" ] && MAIN_START_SECONDS=$(date +%s)
+[ "$debug" ] && MAIN_START_SECONDS=$(date +%s)
 
 main
 
-[ "${debug}" ] && {
+[ "$debug" ] && {
   MAIN_FINISH_SECONDS=$(date +%s)
-  MAIN_ELAPSECS=$(( MAIN_FINISH_SECONDS - MAIN_START_SECONDS ))
-  MAIN_ELAPSED=`eval "echo $(date -ud "@$MAIN_ELAPSECS" +'$((%s/3600/24)) days %H hr %M min %S sec')"`
+  MAIN_ELAPSECS=$((MAIN_FINISH_SECONDS - MAIN_START_SECONDS))
+  MAIN_ELAPSED=$(eval "echo $(date -ud "@$MAIN_ELAPSECS" +'$((%s/3600/24)) days %H hr %M min %S sec')")
   printf "\nTotal elapsed time = %s${MAIN_ELAPSED}\n"
 }
 
-ulimit -n ${currlimit}
+ulimit -n "$currlimit"